tauri-mcp-client/resource/mcp-client-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
pydantic==2.4.2
openai==1.54.3
mcp==1.0.0
